# packages/shared/openapi/spec.yaml
openapi: 3.0.0
info:
  title: Trackfluence API
  description: |
    Revenue Attribution & Intelligence Platform for Creator-Led Growth.

    ## Authentication

    All endpoints (except those marked **Public**) require a JWT Bearer token:

    ```
    Authorization: Bearer <token>
    ```

    Obtain a token via `POST /api/v1/auth/login`.

    ## Rate Limiting

    200 requests per 60 seconds per authenticated user.
    Unauthenticated requests are rate-limited per IP.

    Response headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`

    ## Base URL

    | Environment | URL |
    |-------------|-----|
    | Local dev   | `http://localhost:4000/api/v1` |
    | Production  | `https://api.trackfluence.io/api/v1` |
  version: "1.0.0"
  contact:
    name: Trackfluence Support
    email: support@trackfluence.io
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
  - url: https://api.trackfluence.io/api/v1
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DashboardKPIs:
      type: object
      properties:
        totalRevenue:
          type: number
        attributedRevenue:
          type: number
        attributionRate:
          type: number
        avgOrderValue:
          type: number
        creatorAcquiredCustomers:
          type: integer
        clickToAttributionRate:
          type: number
security:
  - BearerAuth: []
paths:
...
  /revenue-intelligence/dashboard:
    get:
      tags: [Revenue Intelligence]
      summary: Get dashboard KPI metrics
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: KPI metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardKPIs'
...
  /revenue-intelligence/forecast:
    get:
      tags: [Revenue Intelligence]
      summary: Get revenue forecast
      parameters:
        - name: months
          in: query
          schema:
            type: integer
            default: 3
      responses:
        '200':
          description: Forecast data
          content:
            application/json:
              schema:
                type: object
                properties:
                  historical:
                    type: array
                    items:
                      type: object
                      properties:
                        month:
                          type: string
                        revenue:
                          type: number
                  forecast:
                    type: array
                    items:
                      type: object
                      properties:
                        month:
                          type: string
                        projected:
                          type: number
...
  /webhooks:
    get:
      tags: [Webhooks]
      summary: List outbound webhooks
      responses:
        '200':
          description: Webhooks list
    post:
      tags: [Webhooks]
      summary: Create an outbound webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                  description: Leave empty to subscribe to all events
                description:
                  type: string
      responses:
        '201':
          description: Webhook created (secret shown once)
  /webhooks/{id}/toggle:
    post:
      tags: [Webhooks]
      summary: Toggle webhook active/disabled
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated webhook
...